validate CY_TEST_CLUSTER in cypress plugin
