+++ packages/manager/cypress.config.ts
@@ -19,8 +19,11 @@ import {
 } from './cypress/support/plugins/junit-report';
 import { loadEnvironmentConfig } from './cypress/support/plugins/load-env-config';
 import { nodeVersionCheck } from './cypress/support/plugins/node-version-check';
+import {
+  clusterOverrideCheck,
+  regionOverrideCheck,
+} from './cypress/support/plugins/override-check';
 import { postRunCleanup } from './cypress/support/plugins/post-run-cleanup';
-import { regionOverrideCheck } from './cypress/support/plugins/region-override-check';
 import { resetUserPreferences } from './cypress/support/plugins/reset-user-preferences';
 import { splitCypressRun } from './cypress/support/plugins/split-run';
 import { logTestTagInfo } from './cypress/support/plugins/test-tagging-info';
@@ -101,6 +104,7 @@ export default defineConfig({
         fetchLinodeClusters,
         resetUserPreferences,
         regionOverrideCheck,
+        clusterOverrideCheck,
         featureFlagOverrides,
         logTestTagInfo,
         splitCypressRun,
